Refactor imports and improve code structure across multiple modules
- Unified import statements for better readability and organization in `unified_config.py`, `rules.py`, `runner.py`, `jsonl_playback.py`, `jsonl_recording.py`, `aggregation.py`, `extractor_report.py`, `figures.py`, `metadata.py`, `orchestrator.py`, `report_template.py`, `statistics.py`, `robot_state.py`, `fast_pysf_backend.py`, `manifest_writer.py`, `hardware_probe.py`, `multi_extractor_analysis.py`, `multi_extractor_models.py`, and `multi_extractor_paths.py`.
- Replaced direct imports with dynamic imports using `importlib` where applicable to enhance modularity.
- Updated type hints to use the new union syntax (e.g., `int | None`) for better clarity.
- Removed redundant imports and improved exception handling in GPU metadata collection.
- Enhanced the structure of data classes for better type safety and clarity.
- Cleaned up code formatting and ensured consistent use of type hints across the codebase.

+++ robot_sf/benchmark/snqi/cli.py
@@ -130,7 +130,7 @@ def _compute_baseline_stats(episodes: list[dict]) -> dict[str, dict[str, float]]
         values_sorted = sorted(values)
         med = statistics.median(values_sorted)
         # p95 index calculation
-        idx = min(len(values_sorted) - 1, max(0, int(round(0.95 * (len(values_sorted) - 1)))))
+        idx = min(len(values_sorted) - 1, max(0, round(0.95 * (len(values_sorted) - 1))))
         p95 = float(values_sorted[idx])
         if abs(p95 - med) < 1e-12:  # ensure non-zero span for normalization downstream
             p95 = med + 1.0
